Refactor modal UI into feature modules
